Insane version of day 7.
To avoid hash lookups, this version keeps a stack of references into
the dir_size hash table, and uses that to modify the values directly.
Except that to keep the compiler happy, the values in the table are
now Rc<Cell<usize>>>, not <usize>.

This makes no engineering sense, but it wasn't supposed to.  It was
supposed to give me an excuse to play around with internal mutability.